Feature/test repositories

+++ src/main/java/com/ventaproductos/orderitem/entity/OrderItemDTO.java
@@ -2,12 +2,10 @@ package com.ventaproductos.orderitem.entity;
 
 import lombok.Builder;
 
-import java.math.BigDecimal;
-
 @Builder
 public record OrderItemDTO(
-    BigDecimal quantity,
-    BigDecimal unitPrice,
+    int quantity,
+    double unitPrice,
     Integer id,
     Integer orderId,
     Integer productId
