+++ src/main/java/com/ventaproductos/product/entity/ProductDTOSave.java
@@ -1,7 +1,5 @@
 package com.ventaproductos.product.entity;
 
-import java.math.BigDecimal;
-
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 
@@ -11,7 +9,7 @@ public record ProductDTOSave(
     @NotNull(message = "required field")
     String name,
     @NotNull(message = "required field")
-    BigDecimal  price,
+    double  price,
     @NotNull(message = "required field")
     int stock
